requirements.txt
groq==0.9.0
python-dotenv==1.0.0
flask==2.3.3
httpx==0.27.2

